Use double for Shooter Desired RPM

+++ src/main/java/frc/robot/commands/ShooterHoldVelocityCommand.java
@@ -51,7 +51,7 @@ public class ShooterHoldVelocityCommand extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    int desiredRPM;
+    double desiredRPM;
     switch(rpmSource) {
       case CONSTANT:
         desiredRPM = rpm;
